Rename trackApplicationUninstalled to deleteDevice
