+++ packages/destination-actions/src/destinations/customerio/deleteDevice/generated-types.ts
