+++ packages/destination-actions/src/destinations/customerio/deleteDevice/index.ts
